parameterize and split ES spec tests
Currently, it is rather difficult to maintain the ever-growing
elasticsearch spec file. This Commit is an inital refactor of
the spec and splits out some of the specs into their own files,
and separates integration tests from unit tests. Some of the
available options for tests is to set a custom host and port to
test against. These can be set by the `ES_HOST` and `ES_HTTP_PORT`
environment variables.
