+++ README.md
@@ -37,10 +37,22 @@ bundle install
 bundle install
 ```
 
-- Run tests
+- Run unit tests
 
 ```sh
 bundle exec rspec
+```
+
+- Run integration tests
+
+Dependencies: [Docker](http://docker.com)
+
+Before the test suite is run, we will load and run an
+Elasticsearch instance within a docker container. This container 
+will be cleaned up when suite has finished.
+
+```sh
+bundle exec rspec --tag integration
 ```
 
 ### 2. Running your unpublished Plugin in Logstash
