+++ spec/unit/outputs/elasticsearch/protocol_spec.rb
@@ -49,4 +49,3 @@ describe LogStash::Outputs::Elasticsearch::Protocols::HTTPClient do
     end
   end
 end
-
